feat(test): 重整randx目录测试文件结构
将6个测试文件重组为4个标准命名的测试文件：
- 删除孤立文件: benchmark_test.go, coverage_analysis_test.go
- 增强文件: any_test.go, bool_test.go, number_test.go, time_test.go
- 合并基准测试和覆盖率分析到对应源文件的测试中
- 保持99.3%测试覆盖率和91个测试函数
