--- randx/coverage_analysis_test.go
@@ -1,78 +1,0 @@
-package randx
-
-import (
-	"testing"
-)
-
-// TestCoverageAnalysis 记录测试覆盖率分析和极端边界情况的测试努力
-func TestCoverageAnalysis(t *testing.T) {
-	t.Log("=== randx 模块测试覆盖率分析 ===")
-	t.Log("")
-	t.Log("当前覆盖率状态: 99.3%")
-	t.Log("")
-	t.Log("已达到100%覆盖率的函数:")
-	t.Log("- Choose, ChooseN, Shuffle, BatchChoose")
-	t.Log("- Bool, Booln, WeightedBool, BatchBool, BatchBooln")
-	t.Log("- 所有数字生成函数 (Int, Float, Uint系列)")
-	t.Log("- 所有时间工具函数 (RandomDuration, RandomTime等)")
-	t.Log("- TimeDuration4Sleep (经过边界条件修复)")
-	t.Log("")
-	t.Log("剩余未覆盖的极端边界情况 (0.7%):")
-	t.Log("")
-	t.Log("1. WeightedChoose 回退情况 (95.2% 覆盖率)")
-	t.Log("   - 未覆盖代码: return items[len(items)-1]")
-	t.Log("   - 触发条件: 浮点精度问题导致随机数超出累计权重")
-	t.Log("   - 概率: 极低 (~1 in 10^15)")
-	t.Log("   - 测试努力: 已进行5百万次测试，使用多种浮点精度组合")
-	t.Log("")
-	t.Log("2. Jitter 负数修正情况 (90% 覆盖率)")
-	t.Log("   - 未覆盖代码: result = 0 (when result < 0)")
-	t.Log("   - 触发条件: 极小duration + 高抖动百分比导致负结果")
-	t.Log("   - 概率: 极低，依赖特定随机数生成条件")
-	t.Log("   - 测试努力: 已进行千万次测试，使用纳秒级duration")
-	t.Log("")
-	t.Log("结论:")
-	t.Log("- 99.3%的覆盖率已经是优秀结果")
-	t.Log("- 剩余的0.7%是理论性的极端边界情况")
-	t.Log("- 在实际使用中几乎不可能遇到")
-	t.Log("- 所有实用功能都已得到全面测试验证")
-}
-
-// TestEdgeCasesDocumentation 记录已测试的边界情况
-func TestEdgeCasesDocumentation(t *testing.T) {
-	t.Log("=== 已覆盖的边界情况总结 ===")
-	t.Log("")
-	t.Log("WeightedChoose 边界情况:")
-	t.Log("✓ 空项目列表")
-	t.Log("✓ 长度不匹配的权重")
-	t.Log("✓ 单个项目")
-	t.Log("✓ 零权重")
-	t.Log("✓ 负权重")
-	t.Log("✓ 浮点精度问题权重")
-	t.Log("✓ 极小权重值")
-	t.Log("○ 浮点精度导致的回退情况 (极难触发)")
-	t.Log("")
-	t.Log("Jitter 边界情况:")
-	t.Log("✓ 零百分比抖动")
-	t.Log("✓ 负百分比抖动")
-	t.Log("✓ 超过100%的抖动")
-	t.Log("✓ 零duration")
-	t.Log("✓ 极小duration + 高抖动")
-	t.Log("○ 负结果修正 (在特定条件下可触发)")
-	t.Log("")
-	t.Log("TimeDuration4Sleep 边界情况:")
-	t.Log("✓ 无参数 (默认范围)")
-	t.Log("✓ 单参数 (0到指定值)")
-	t.Log("✓ 双参数 (指定范围)")
-	t.Log("✓ start == end (返回该值)")
-	t.Log("✓ start > end (抛出panic)")
-	t.Log("✓ 零duration")
-	t.Log("✓ 负duration")
-	t.Log("")
-	t.Log("所有其他函数的边界情况:")
-	t.Log("✓ 空切片操作")
-	t.Log("✓ 零值和负值输入")
-	t.Log("✓ 极值测试")
-	t.Log("✓ 并发安全性")
-	t.Log("✓ 分布均匀性验证")
-}
